added bannecomands

+++ OilChangeApp/Domain/Entities/BanedUsers.cs
@@ -3,7 +3,7 @@ namespace OilChangeApp.Domain.Entities;
 public class BanedUsers
 {
     [Key]
-    public int BanedTgId { get; set; }
+    public long BanedTgId { get; set; }
     [Column("ExpiredDate")]
     public DateTime ExpiredDate { get; set; }
 }
